AnchoredVWAP with VWAPBase class
